addtional clean up of the info
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@28589 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Queries/Queries/avtIntegralCurveInfoQuery.h
@@ -88,9 +88,10 @@ class QUERY_API avtIntegralCurveInfoQuery : public avtDatasetQuery
     virtual void             PostExecute(void);
     virtual void             Execute(vtkDataSet *ds, const int chunk);
 
-    int  dumpOpts;
+    unsigned int dumpOpts;
     bool dumpValues;
-    std::vector<float> slData;
+
+    std::vector<double> slData;
 };
 
 #endif
